root admin spike

+++ Gateway/src/admin.ts
@@ -1,6 +1,7 @@
 import { Router, Request, Response } from "express";
 import { Storage, UserModel, UserUpdateModel, userViewModel, UserViewModel } from "./storage";
 import { badRequest, okResult } from "./utils";
+import { rootAdminName } from "./root-admin";
 
 export const admin = (storage: Storage) =>
     Router()
@@ -11,6 +12,11 @@ export const admin = (storage: Storage) =>
         })
         .put("/users/:username", async (req: Request, res: Response) => {
             const username: string = (req.params as any).username;
+
+            if (username === rootAdminName) {
+                badRequest(res, "it is forbidden to edit root admin");
+                return;
+            }
 
             const updateModel: UserUpdateModel | undefined = (req as any).body.userUpdate;
             if (!updateModel) {
@@ -34,6 +40,11 @@ export const admin = (storage: Storage) =>
         })
         .delete("/users/:username", async (req: Request, res: Response) => {
             const username: string = (req.params as any).username;
+
+            if (username === rootAdminName) {
+                badRequest(res, "it is forbidden to delete root admin");
+                return;
+            }
 
             const success = await storage.deleteUser(username);
             if (!success) {
